+++ examples/basic/basic.js
@@ -2,34 +2,7 @@ import React, { Component } from "react";
 import ReactDOM from "react-dom";
 import DateTimeField from "react-bootstrap-datetimepicker";
 import moment from "moment";
-
-class ParentComponent extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      date: "1990-06-05",
-      format: "YYYY-MM-DD",
-      inputFormat: "DD/MM/YYYY",
-      mode: "date"
-    };
-  }
-
-  handleChange = (newDate) => {
-    console.log("newDate", newDate);
-    return this.setState({date: newDate});
-  }
-
-  render() {
-    const {date, format, mode, inputFormat} = this.state;
-    return <DateTimeField
-      dateTime={date}
-      format={format}
-      viewMode={mode}
-      inputFormat={inputFormat}
-      onChange={this.handleChange}
-    />;
-  }
-}
+import ParentComponent from "./ParentComponent";
 
 class Basic extends Component {
 
@@ -155,7 +128,5 @@ class Basic extends Component {
       );
    }
 }
-
-
 
 ReactDOM.render(React.createFactory(Basic)(), document.getElementById("example"));
